pyproject.toml
[project]
name = "octapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "pyqt5>=5.15.11",
]

[dependency-groups]
dev = [
    "pillow>=12.0.0",
    "pytest>=9.0.2",
    "pytest-flask>=1.3.0",
]
